Migrate aggregate
